Deferred operand stack and local var table saving
Deferred operand stack and local variable table saving until Coroutine
suspended. Previously, we saved the entire operand stack and local
variable table prior to calling in to a method that took in a Continuation
object. Now, we only save a tiny bit of the operand stack (just the args
for the method being invoked) and if we see that we've entered in to a
saving state once the method we invoked returns, we save the rest of it.

All in all, for any non-trivial application, this will be a major
performance boost.