Reload locals if debugMode is true
This change always reloads the local variables table if debugMode is set
to true, even if it isn't needed. The reason for this is that if you set a
breakpoint in your method and you want to inspect the local variables
anywhere in the call chain, we need to make sure the local variables are
always loaded for everything when the coroutine resumes.