Optimize pushing/popping of method states
Previously, the method state stack was fully being popped off during a
restore and fully being pushed back on after a suspend. This is
inefficient, especially when we have large call depths. As such, this was
refactored to use pointers rather than pushing/popping off the stack.

The new code will move the pointer all the way down to the last method
state, and then move it back up as the restored methods finish executing.
Once the coroutine suspends again only the method states for the new
invocation are tacked on.

What this is doing right now is incredibly convoluted. It may not provide
a 'massive' performance gain unless you have a low-IO high-compute
coroutine. BE PREPARED TO REMOVE THIS IN THE FUTURE.